feat: add workflow note
